+++ src/tests/server.ts
@@ -1,0 +1,15 @@
+import { setupServer } from 'msw/node';
+import { afterAll, afterEach, beforeAll } from 'vitest';
+
+export const restHandlers = [];
+
+export const server = setupServer(...restHandlers);
+
+// Start server before all tests
+beforeAll(() => server.listen({ onUnhandledRequest: 'error' }));
+
+// Close server after all tests
+afterAll(() => server.close());
+
+// Reset handlers after each test for test isolation
+afterEach(() => server.resetHandlers());
